Ft/ingredient crud (#57)
* Added api endpoints for ingredients

* Added configurable allowOrigin to cors middleware

* Changed Gin to Release Mode for Production

* fixed issue with getting userId parameter for adding ingredient

* fixed userID for listing ingredients

* used VM for ingredient listing api to hide userId

* changed max ingredients per page to 20

* removed public aws login

* Added to db architecture with corresponding Sqlc model templates.

Updated tests to reflect these changes

* Fixed conversion issue with inci value in ingredients

* fixed unit test using new inci type

* Added Formula api endpoints, models and utility functions

* Cleaned up ingress in prep for tls

* activated allow origin for temporary front-end

* reset OPTIONS response from defaulting to 204

* Specified allow-origin to staged front-end

* Added 204 response to OPTIONS requests

* Added refresh sessions with tokens and api endpoints

* Added migrations for sessions

* Removed uuid dependency

* Removed Inci from required for adding Ingredients

* Added count, update and delete Ingredient endpoints and functionality.

* Changed session id to uuid

* Added import for uuid in deploy

* Added import for uuid in deploy in test

* reset moved tidy in test

* reset moved tidy in deploy

* Added support for ingredient tags

prepared migrations for formula tags

* Added support for ingredient tags

prepared migrations for formula tags
removed old relation in migration

* Added support for ingredient tags

prepared migrations for formula tags
removed old relation in migration
fixed ingredient api test

* Added cost to ingredient table

Added cost to ingredient update endpoint
Fixed issue with capitalized tags not being saved

* Added cost to listIngredient response

* Added Cost field to Formula Ingredient.

* Fixed issue where removing phases and ingredients as well as adding phases was not reflected in db

Made Listing Phases through GetFullFormula to order by phaseId to put the first one created on top, this is how it should behave on front-ends stack-like adding.

* Increased max ingredient listing to 250 and made order alphabetically

* Enabled cost and percentages for formula ingredients to be floats.
Ordering phases by created_at

* Enabled cost and percentages for formula ingredients to be floats.
Ordering phases by created_at

* Updated formulas test to reflect int to float changes in db

+++ api/ingredients.go
@@ -184,7 +184,7 @@ func (server *Server) updateIngredient(ctx *gin.Context) {
 		Name:       req.Name,
 		Inci:       req.Inci,
 		Hash:       "",
-		Cost:       sql.NullInt32{Int32: int32(req.Cost), Valid: true},
+		Cost:       sql.NullFloat64{Float64: float64(req.Cost), Valid: true},
 		UserID:     user.ID,
 		FunctionID: sql.NullInt64{},
 	}
@@ -309,7 +309,7 @@ func newIngredientResponse(ingredient db.Ingredient, tags []ingredientTagRequest
 		Id:   ingredient.ID,
 		Name: ingredient.Name,
 		Inci: ingredient.Inci,
-		Cost: ingredient.Cost.Int32,
+		Cost: int32(ingredient.Cost.Float64),
 		Tags: tags,
 	}
 }
